[refactor] - Logout from Profile
