+++ src/Pages/CSS/profile.css
@@ -14,3 +14,11 @@
     border-bottom: 1px solid;
     padding: 1rem 1rem 0.5rem 1rem;
 }
+
+.tab:hover{
+    color: var(--accent-color);
+}
+
+.tab:active{
+    color: var(--accent-color);
+}
